pyproject.toml
[project]
name = "sprintlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "numpy>=2.3.2",
    "omegaconf>=2.3.0",
    "rich>=14.1.0",
    "tiktoken>=0.9.0",
    "torch>=2.7.1",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.5",
    "mkdocs-material>=9.6.16",
    "plotly>=6.2.0",
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
